Split datahub.company.urls into subpackage
