+++ config/api_urls.py
@@ -4,8 +4,10 @@ from django.urls import include, path
 from rest_framework import routers
 
 from datahub.activity_stream import urls as activity_stream_urls
-from datahub.company import urls as company_urls
 from datahub.company import views as company_views
+from datahub.company.urls import ch_company as ch_company_urls
+from datahub.company.urls import company as company_urls
+from datahub.company.urls import contact as contact_urls
 from datahub.event import urls as event_urls
 from datahub.feature_flag import urls as feature_flag_urls
 from datahub.interaction import urls as interaction_urls
@@ -32,9 +34,9 @@ v3_urls = [
             namespace='activity-stream',
         ),
     ),
-    path('', include((company_urls.contact_urls, 'contact'), namespace='contact')),
-    path('', include((company_urls.company_urls_v3, 'company'), namespace='company')),
-    path('', include((company_urls.ch_company_urls_v3, 'ch-company'), namespace='ch-company')),
+    path('', include((ch_company_urls.urls_v3, 'ch-company'), namespace='ch-company')),
+    path('', include((company_urls.urls_v3, 'company'), namespace='company')),
+    path('', include((contact_urls.urls_v3, 'contact'), namespace='contact')),
     path('', include((event_urls, 'event'), namespace='event')),
     path('', include((feature_flag_urls, 'feature-flag'), namespace='feature-flag')),
     path('', include((interaction_urls, 'interaction'), namespace='interaction')),
@@ -54,7 +56,7 @@ v3_urls = [
 # API V4 - new format for addresses
 
 v4_urls = [
-    path('', include((company_urls.company_urls_v4, 'company'), namespace='company')),
-    path('', include((company_urls.ch_company_urls_v4, 'ch-company'), namespace='ch-company')),
+    path('', include((ch_company_urls.urls_v4, 'ch-company'), namespace='ch-company')),
+    path('', include((company_urls.urls_v4, 'company'), namespace='company')),
     path('', include((search_urls.urls_v4, 'search'), namespace='search')),
 ]
